Text.EDE

Description

A (mostly logicless) textual templating language in the same vein as Liquid or Jinja2.

(ED-E is a character from Fallout New Vegas, pronounced iː-diː-iː, or Eddie.)

How to use this library

A simple example of parsing and rendering Text containing a basic conditional expression and variable interpolation follows.

First the Template is defined:

>>> let tmpl = parse "{% if var %}Hello, {{ var }}!{% else %}negative!{% endif %}" :: Result Template

Then an Object is defined containing the environment which will be available to the Template during rendering:

>>> let env = render $ fromPairs [ "var" .= "World" ] :: Template -> Result Text

Finally the environment is applied to the Template:

>>> tmpl >>= env :: Result Text
> Success "Hello, World!"

In this manner, Templates can be pre-compiled to the internal AST and the cost of parsing can be amortised if the same Template is rendered multiple times.

Another example, this time rendering a Template from a file:

 import qualified Data.Text.Lazy as LText
 import           Text.EDE

 main :: IO ()
 main = do
     tmpl <- LText.readFile "template.ede"
     either error print $ eitherParse tmpl >>= eitherRender env
   where
     env = fromPairs
         [ "text" .= "Some Text."
         , "int"  .= 1
         , "list" .= [5..10]
         ]

Please see the syntax section for more information about available statements and expressions.

Parsing and Rendering

Parsing and rendering require two separate steps intentionally so that the more expensive result of parsing can be embedded, and/or reused.

Parsing tokenises the input and converts it to an internal AST representation, which can be cached for future use.

Rendering takes an Object as the environment and a parsed Template to subsitute the values into.

data Template

A valid parsed and compiled template.

Basic

parse

Arguments

:: Text
Lazy Text template definition.

-> Result Template

Parse Text into a compiled template.

This function is pure and does not resolve include expressions. This requires that the caller supplies a HashMap of Templates at render time using renderWith.

render :: Template -> Object -> Result Text

Render an Object using the supplied Template.

Rendering a Template which has been been parsed used the pure variants of parse, and contains

IO and Includes

parseFile

Arguments

:: Bool
Error if target include tag is a non-existent file.

-> FilePath
Path to the template to load.

-> IO (Result Template)

Read and parse a file into a compiled template.

If True is specified as the first parameter, parseFile performs all template resolution and parsing recursively, with non-existent templates and parse errors returning an error immediately.

If you wish to defer loading of includes, or supply additional precompiled/named includes at runtime (using renderWith rather than exclusively using file paths) you should specify False as the first parameter, or use parse and renderWith directly.

Configurable

parseAs

Arguments

:: SourceName
Descriptive name for error messages.

-> Text
Lazy Text template definition.

-> Result Template

Parse Text into a compiled template.

renderWith

Arguments

:: HashMap Text Fun
Filters

-> HashMap Text Template
Additional Templates

-> Template
Parsed Template

-> Object
Environment

-> Result Text

Render an Object using the supplied Template.

Either Variants

eitherParse :: Text -> Either String Template

See: parse

eitherParseFile :: Bool -> FilePath -> IO (Either String Template)

See: parseFile

eitherParseAs :: SourceName -> Text -> Either String Template

See: parseAs

eitherRender :: Template -> Object -> Either String Text

See: render

eitherRenderWith :: HashMap Text Fun -> HashMap Text Template -> Template -> Object -> Either String Text

See: renderWith

Results and Errors

Unsuccessful parse or render steps can be inspected or analysed using result.

>>> result failure success $ parse tmpl >>= render env

If you're only interested in dealing with errors as strings, and the positional information contained in Meta is not of use you can use the convenience functions eitherParse, eitherRender, or convert a Result to Either using eitherResult.

>>> either failure success $ eitherParse tmpl >>= eitherRender env

data Meta

Meta information describing the source position of an expression or error.

Constructors

Meta

Fields

metaSource :: !String

metaRow :: !Int

metaColumn :: !Int

data Result a

The result of running parsing or rendering steps.

Constructors

Error !Meta [String]

Success a

eitherResult :: Result a -> Either String a

Convert a Result to an Either with the Left case holding a formatted error message, and Right being the successful result over which Result is paramterised.

result

Arguments

:: (Meta -> [String] -> b)
Function to apply to the Error parameters.

-> (a -> b)
Function to apply to the Success case.

-> Result a
The Result to map over.

-> b

Perform a case analysis on a Result.

Includes

Filters

data Fun where

Constructors

Fun :: (Eq a, Eq b) => TType a -> TType b -> (a -> b) -> Fun

data TType a where

Constructors

TNil :: TType ()

TText :: TType Text

TBool :: TType Bool

TInt :: TType Integer

TDbl :: TType Double

TBld :: TType Builder

TMap :: TType Object

TList :: TType Array

TFun :: TType Fun

defaultFilters :: HashMap Text Fun

Convenience

Data.Aeson

fromPairs is a wrapper around Aeson's object function which safely strips the outer Value constructor, providing the correct type signature for input into render.

It is used in combination with the re-exported .= as follows:

>>> render (fromPairs [ "foo" .= "value", "bar" .= 1 ]) :: Template -> Result Text

fromPairs :: [Pair] -> Object

Create an Object from a list of name/value Pairs. See Aeson's documentation for more details.

(.=) :: ToJSON a => Text -> a -> Pair

Syntax

Comments

Comments are ignored by the parser and omitted from the rendered output.

 {# singleline comment #}

 {#
    multiline
    comment
 #}

Raw

You can disable template processing for blocks of text using the raw section:

 {% raw %}
 Some {{{ handlebars }}} or {{ mustache }} or {{ jinja2 }} output tags etc.
 {% endraw %}

This can be used to avoid parsing expressions which would otherwise be considered valid ED-E syntax.

Variables

Variables are substituted directly for their Buildable representation. An error is raised if the varaible being substituted is not a literal type (ie. an Array or Object) or doesn't exist in the supplied environment.

 {{ var }}

Nested variable access is also supported for variables which resolve to an Object. Dot delimiters are used to chain access through multiple nested Objects. The right-most accessor must resolve to a Buildable type as with the previous non-nested variable access.

 {{ nested.var.access }}

Conditionals

A conditional is introduced and completed with the section syntax:

 {% if var %}
    ... consequent expressions
 {% else %}
    ... alternate expressions
 {% endif %}

The value of {{ var }} determines the branch that is rendered by the template with the else branch being optional.

In the case of a literal it conforms directly to the supported boolean or relation logical operators from Haskell. If a variable is singuarly used its existence determines the result of the predicate, the exception to this rule is boolean values which will be substituted into the expression if they exist in the supplied environment.

The following logical expressions are supported as predicates in conditional statements with parameters being type checked and an error is raised if the left and right hand sides are not type equivalent.

And: &&

Or: ||

Equal: ==

Not Equal: /=

Greater: >

Greater Or Equal: >=

Less: <

Less Or Equal: <=

Negation: !

Case Analysis

To pattern match a literal or variable, you can use the case statement:

 {% case var %}
 {% when "a" %}
    .. matched expressions
 {% when "b" %}
    .. matched expressions
 {% else %}
    .. alternate expressions
 {% endcase %}

Loops

Iterating over an Array or Object can be acheived using the 'for ... in' section syntax. Attempting to iterate over any other type will raise an error.

Example:

 {% for var in list %}
     ... iteration expression
 {% else %}
     ... alternate expression
 {% endfor %}

The iteration branch is rendering per item with the else branch being (which is optional) being rendered if the {{ list }} variable is empty.

When iterating over an Object, a stable sort using key equivalence is applied, Arrays are unmodified.

The resulting binding within the iteration expression (in this case, {{ var }}) is an Object containing the following keys:

key :: Text: They key if the loop target is an Object

value :: a: The value of the loop target

loop :: Object: Loop metadata.

length :: Int: Length of the loop

index :: Int: Index of the iteration

index0 :: Int: Zero based index of the iteration

remainder :: Int: Remaining number of iterations

remainder0 :: Int: Zero based remaining number of iterations

first :: Bool: Is this the first iteration?

last :: Bool: Is this the last iteration?

odd :: Bool: Is this an odd iteration?

even :: Bool: Is this an even iteration?

For example:

 {% for item in items %}
     {{ item.index }}:{{ item.value }}
     {% if !item.last %}

     {% endif %}
 {% endfor %}

Will render each item with its (1-based) loop index as a prefix, separated by a blank newline, without a trailing at the end of the document.

Includes

TODO

Filters

Filters are typed functions which can be applied to variables and literals. An example of rendering a lower cased boolean would be:

 {{ True | show | lower }}

The input is on the LHS and chained filters (delimited by '|') are on the RHS, with filters being applied left associatively.

Available filters:

show :: a -> Text: Convert a value to Text using its Show instance.

lower :: Text -> Text: Lower case a textual value.

upper :: Text -> Text: Upper case a textual value.