Text.EDE

Description

A (mostly logicless) textual templating language in the same vein as Liquid or Jinja2.

(ED-E is a character from Fallout New Vegas, pronounced iː-diː-iː, or Eddie.)

How to use this library

A simple example of parsing and rendering Text containing a basic conditional expression and variable interpolation follows.

First the Template is defined:

 >>> let tmpl = parse "{% if var %}Hello, {{ var }}!{% else %}negative!{% endif %}" :: Result Template

Then an Object is defined containing the environment which will be available to the Template during rendering:

 >>> let env = render $ fromPairs [ "var" .= "World" ] :: Template -> Result Builder

Finally the environment is applied to the Template:

 >>> tmpl >>= env :: Result Builder
 Success "Hello, World!"

In this manner, Templates can be pre-compiled to the internal AST and the cost of parsing can be amortised if the same Template is rendered multiple times.

Parsing and rendering

Parsing and rendering require two separate steps intentionally so that the more expensive result of parsing can be embedded, and/or reused.

Parsing tokenises the input and converts it to an internal AST representation, which can be cached for future use.

Rendering takes an Object as the environment and a parsed Template to subsitute the values into.

data Template

A valid parsed and compiled template.

parse :: Text -> Result Template

Parse Text into a compiled template.

render :: Object -> Template -> Result Builder

Render an Object using the supplied Template.

eitherParse :: Text -> Either String Template

Parse Text into a compiled template, and convert the Result using eitherResult.

eitherRender :: Object -> Template -> Either String Builder

Render an Object using the supplied Template, and convert the Result using eitherResult.

Results and errors

Unsuccessful parse or render steps can be inspected or analysed using result.

 >>> result failure success $ parse tmpl >>= render env

If you're only interested in dealing with errors as strings, and the positional information contained in Meta is not of use you can use the convenience functions eitherParse, eitherRender, or convert a Result to Either using eitherResult.

 >>> either failure success $ eitherParse tmpl >>= eitherRender env

data Meta

Meta information describing the source position of an expression or error.

Constructors

Meta

Fields

_source :: !String

_row :: !Int

_column :: !Int

data Result a

The result of running parsing or rendering steps.

Constructors

Error Meta [String]

Success a

eitherResult :: Result a -> Either String a

Convert a Result to an Either with the Left case holding a formatted error message, and Right being the successful result over which Result is paramterised.

result

Arguments

:: (Meta -> [String] -> b)
Function to apply to the Error parameters.

-> (a -> b)
Function to apply to the Success case.

-> Result a
The Result to map over.

-> b

Perform a case analysis on a Result.

Convenience

Data.Aeson

fromPairs is a wrapper around Aeson's object function which safely strips the outer Value constructor, providing the correct type signature for input into render.

It is used in combination with the re-exported .= as follows:

 >>> render (fromPairs [ "foo" .= "value", "bar" .= 1 ]) :: Template -> Result Builder

fromPairs :: [Pair] -> Object

Create an Object from a list of name/value Pairs. See Aeson's documentation for more details.

(.=) :: ToJSON a => Text -> a -> Pair

Data.Text.Lazy.Builder

The successful result of rendering an Object environment and Template is a lazy Builder which can be converted to Text using the re-exported toLazyText.

toLazyText :: Builder -> Text

Syntax

Comments

Comments are ignored by the parser and omitted from the rendered output.

 {# singleline comment #}

 {#
    multiline
    comment
 #}

Variables

Variables are substituted directly for their Buildable representation. An error is raised if the varaible being substituted is not a literal type (ie. an Array or Object).

 {{ var }}

Nested variable access is also supported for variables which resolve to an Object. Dot delimiters are used to chain access through multiple nested Objects. The right-most accessor must resolve to a Buildable type as with the previous non-nested variable access.

 {{ nested.var.access }}

Conditionals

Boolean Logic

Relational Logic

Negation

Case Statements

Loops

Context

Debugging